Remove previous iteration of dice, rename DieConstants
Doesn't make sense to have it be a constants file anymore, so now it is
the premier Dice file!
